requirements/testing.txt
-r common.txt

django-discover-runner==0.2
factory-boy==1.1.4